refactor folder structure on watch target
