+++ BookPlayerWatch/LocalPlayback/Player/PlayerToolbarView.swift
@@ -9,49 +9,6 @@
 import BookPlayerWatchKit
 import SwiftUI
 
-final class PlaybackFullControlsViewModel: ObservableObject {
-  let playerManager: PlayerManager
-
-  var rate: Float {
-    self.playerManager.currentSpeed
-  }
-
-  var boostVolume: Bool {
-    UserDefaults.standard.bool(forKey: Constants.UserDefaults.boostVolumeEnabled)
-  }
-
-  init(playerManager: PlayerManager) {
-    self.playerManager = playerManager
-  }
-
-  func handleBoostVolumeToggle(_ flag: Bool) {
-    self.playerManager.setBoostVolume(flag)
-  }
-
-  func handleNewSpeed(_ rate: Float) {
-    let roundedValue = round(rate * 100) / 100.0
-
-    guard roundedValue >= 0.5 && roundedValue <= 4.0 else { return }
-
-    self.playerManager.setSpeed(roundedValue)
-  }
-
-  func handleNewSpeedJump() {
-    let rate: Float
-
-    if self.rate == 4.0 {
-      rate = 0.5
-    } else {
-      rate = min(self.rate + 0.5, 4.0)
-    }
-
-    let roundedValue = round(rate * 100) / 100.0
-
-    self.playerManager.setSpeed(roundedValue)
-  }
-
-}
-
 struct PlayerToolbarView: View {
   /// PlayerManager needs to be an ObservedObject, otherwise the Published property do not reload properly
   @ObservedObject var playerManager: PlayerManager
@@ -69,7 +26,7 @@ struct PlayerToolbarView: View {
       Spacer()
 
       NavigationLink(
-        destination: PlaybackFullControlsView(model: PlaybackFullControlsViewModel(playerManager: playerManager))
+        destination: PlaybackFullControlsView(model: .init(playerManager: playerManager))
       ) {
         ResizeableImageView(name: "dial.max")
           .accessibilityLabel("settings_controls_title".localized)
